refactor(WinningLottery) : 밸리데이션 체크 및 생성자 로직 위치 변경
* InputView에서 당첨 로또에 대한 밸리데이션 체크를 진행하는 것이 옳지 않다고 생각하여 진행
