+++ src/main/kotlin/lotto/view/InputView.kt
@@ -16,14 +16,9 @@ object InputView {
         return userInput.toInt()
     }
 
-    fun inputWinningNum(): List<Int> {
+    fun inputWinningNum(): String {
         println(INPUT_WINNING_NUMBER_TEXT)
-        val userInput = validateWinningNum(Console.readLine())
-        val numbers = mutableListOf<Int>()
-        userInput.forEach {
-            numbers.add(it.toInt())
-        }
-        return numbers
+        return Console.readLine()
     }
 
     fun inputBonusNum(): Int {
@@ -42,18 +37,7 @@ object InputView {
             "$ERROR_MESSAGE 1,000원 단위로 입력해 주세요."
         }
     }
-
-    private fun validateWinningNum(userInput: String): List<String> {
-        val splitedNum = userInput.split(',')
-        val numRegex = Regex("[^0-9]")
-        require(userInput.isNotBlank() && splitedNum.joinToString("").all { it.isDigit() }) {
-            "$ERROR_MESSAGE 로또 번호는 숫자로 입력해 주세요."
-        }
-        require(splitedNum.all { it.toInt() in 1..45 }) {
-            "$ERROR_MESSAGE 로또 번호는 1부터 45 사이의 숫자여야 합니다."
-        }
-        return splitedNum
-    }
+    
 
     private fun validateBonusNum(userInput: String) {
         require(userInput.isNotBlank() &&
